IOS-3827 Update mentions texts if needed
